# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Video Integration Runtime",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/settings/media-tools": {
      "put": {
        "tags": [
          "T02 media"
        ],
        "summary": "Configure",
        "operationId": "configureMediaTools",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Command_ConfigureTools_"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_MutationReceipt_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/imports": {
      "post": {
        "tags": [
          "T02 media"
        ],
        "summary": "Import Media",
        "operationId": "importMedia",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Command_ImportMedia_"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_MutationReceipt_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/media": {
      "get": {
        "tags": [
          "T02 media"
        ],
        "summary": "List Media",
        "operationId": "listMedia",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_MediaPage_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/media/{media_id}/metadata": {
      "get": {
        "tags": [
          "T02 media"
        ],
        "summary": "Metadata",
        "operationId": "getMediaMetadata",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "media_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
              "format": "uuid",
              "title": "Media Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Media_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/media/{media_id}": {
      "head": {
        "tags": [
          "T02 media"
        ],
        "summary": "Binary",
        "operationId": "headMediaBytes",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "media_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
              "format": "uuid",
              "title": "Media Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "T02 media"
        ],
        "summary": "Binary",
        "operationId": "getMediaBytes",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "media_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
              "format": "uuid",
              "title": "Media Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/media/{media_id}/relocate": {
      "post": {
        "tags": [
          "T02 media"
        ],
        "summary": "Relocate",
        "operationId": "relocateMedia",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "media_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
              "format": "uuid",
              "title": "Media Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Command_RelocateMedia_"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_MutationReceipt_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/jobs": {
      "get": {
        "tags": [
          "T02 media"
        ],
        "summary": "Jobs",
        "operationId": "listLocalJobs",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_list_Job__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/jobs/{job_id}": {
      "get": {
        "tags": [
          "T02 media"
        ],
        "summary": "Job",
        "operationId": "getLocalJob",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Job_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/jobs/{job_id}/cancel": {
      "post": {
        "tags": [
          "T02 media"
        ],
        "summary": "Cancel",
        "operationId": "cancelLocalJob",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Command_CancelJob_"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_MutationReceipt_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CancelJob": {
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "CancelJob"
      },
...
      "Command_CancelJob_": {
        "properties": {
          "clientOperationId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "expectedRevision": {
            "$ref": "#/components/schemas/SafeInteger"
          },
          "payload": {
            "$ref": "#/components/schemas/CancelJob"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "clientOperationId",
          "expectedRevision",
          "payload"
        ],
        "title": "Command[CancelJob]"
      },
...
      "Command_ConfigureTools_": {
        "properties": {
          "clientOperationId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "expectedRevision": {
            "$ref": "#/components/schemas/SafeInteger"
          },
          "payload": {
            "$ref": "#/components/schemas/ConfigureTools"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "clientOperationId",
          "expectedRevision",
          "payload"
        ],
        "title": "Command[ConfigureTools]"
      },
...
      "Command_ImportMedia_": {
        "properties": {
          "clientOperationId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "expectedRevision": {
            "$ref": "#/components/schemas/SafeInteger"
          },
          "payload": {
            "$ref": "#/components/schemas/ImportMedia"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "clientOperationId",
          "expectedRevision",
          "payload"
        ],
        "title": "Command[ImportMedia]"
      },
...
      "Command_RelocateMedia_": {
        "properties": {
          "clientOperationId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "expectedRevision": {
            "$ref": "#/components/schemas/SafeInteger"
          },
          "payload": {
            "$ref": "#/components/schemas/RelocateMedia"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "clientOperationId",
          "expectedRevision",
          "payload"
        ],
        "title": "Command[RelocateMedia]"
      },
...
      "ConfigureTools": {
        "properties": {
          "ffmpegGrantId": {
            "$ref": "#/components/schemas/Uuid"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "ffmpegGrantId"
        ],
        "title": "ConfigureTools"
      },
...
      "Envelope_Job_": {
        "properties": {
          "requestId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "data": {
            "$ref": "#/components/schemas/Job"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "requestId",
          "data"
        ],
        "title": "Envelope[Job]"
      },
      "Envelope_MediaPage_": {
        "properties": {
          "requestId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "data": {
            "$ref": "#/components/schemas/MediaPage"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "requestId",
          "data"
        ],
        "title": "Envelope[MediaPage]"
      },
      "Envelope_Media_": {
        "properties": {
          "requestId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "data": {
            "$ref": "#/components/schemas/Media"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "requestId",
          "data"
        ],
        "title": "Envelope[Media]"
      },
      "Envelope_MutationReceipt_": {
        "properties": {
          "requestId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "data": {
            "$ref": "#/components/schemas/MutationReceipt"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "requestId",
          "data"
        ],
        "title": "Envelope[MutationReceipt]"
      },
...
      "Envelope_list_Job__": {
        "properties": {
          "requestId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/Job"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "requestId",
          "data"
        ],
        "title": "Envelope[list[Job]]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImportMedia": {
        "properties": {
          "fileGrantId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "purpose": {
            "type": "string",
            "enum": [
              "reference",
              "speech",
              "video",
              "music",
              "sfx",
              "evidence"
            ],
            "title": "Purpose"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "fileGrantId",
          "purpose"
        ],
        "title": "ImportMedia"
      },
...
      "Media": {
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Uuid"
          },
          "sha256": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$",
            "title": "Sha256"
          },
          "byteLength": {
            "$ref": "#/components/schemas/PositiveInteger"
          },
          "mime": {
            "type": "string",
            "enum": [
              "image/png",
              "image/jpeg",
              "video/mp4",
              "audio/wav",
              "audio/mpeg",
              "audio/mp4"
            ],
            "title": "Mime"
          },
          "durationMs": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PositiveInteger"
              },
              {
                "type": "null"
              }
            ]
          },
          "width": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PositiveInteger"
              },
              {
                "type": "null"
              }
            ]
          },
          "height": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PositiveInteger"
              },
              {
                "type": "null"
              }
            ]
          },
          "availability": {
            "type": "string",
            "enum": [
              "staging",
              "available",
              "missing",
              "quarantined"
            ],
            "title": "Availability"
          },
          "provenance": {
            "type": "string",
            "enum": [
              "imported",
              "generated",
              "derived",
              "synthetic"
            ],
            "title": "Provenance"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "sha256",
          "byteLength",
          "mime",
          "durationMs",
          "width",
          "height",
          "availability",
          "provenance"
        ],
        "title": "Media"
      },
      "MediaPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Media"
            },
            "type": "array",
            "maxItems": 200,
            "title": "Items"
          },
          "nextCursor": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Uuid"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "nextCursor"
        ],
        "title": "MediaPage"
      },
      "MutationReceipt": {
        "properties": {
          "operationId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "committedRevision": {
            "$ref": "#/components/schemas/SafeInteger"
          },
          "resourceId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "state": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "State"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "operationId",
          "committedRevision",
          "resourceId",
          "state"
        ],
        "title": "MutationReceipt"
      },
...
      "PositiveInteger": {
        "type": "integer",
        "ge": 1,
        "le": 9007199254740991
      },
...
      "RelocateMedia": {
        "properties": {
          "fileGrantId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "expectedHash": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$",
            "title": "Expectedhash"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "fileGrantId",
          "expectedHash"
        ],
        "title": "RelocateMedia"
      },
...
      "SafeInteger": {
        "type": "integer",
        "ge": 0,
        "le": 9007199254740991
      },
...
      "Uuid": {
        "type": "string",
        "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
        "format": "uuid"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "RuntimeBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "32-byte random token, base64url without padding"
      }
    }
  },
  "security": [
    {
      "RuntimeBearer": []
    }
  ]
}